add e2e
